requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
Flask-WTF==1.0.0
email-validator==1.1.3
selenium==4.1.0
webdriver-manager==3.5.2
pywhatkit==5.4
